Refactor: 변수 인자 Long에서 int로 변경

+++ Numble-Challenge/src/main/java/Numble/NumbleChallenge/domain/order/service/OrderService.java
@@ -5,7 +5,7 @@ import Numble.NumbleChallenge.domain.order.Order;
 import java.util.List;
 
 public interface OrderService {
-    Order createOrder(Long memberId, Long itemId, Long count);
+    Order createOrder(Long memberId, Long itemId, int count);
     void cancelOrder(Long orderId);
     List<Order> findAllOrder();
